+++ allure-notifications-api/src/main/java/guru/qa/allure/notifications/model/phrases/Phrases.java
@@ -18,6 +18,8 @@ public class Phrases {
     private String comment;
     @SerializedName("reportAvailableAtLink")
     private String reportAvailableAtLink;
+    @SerializedName("links")
+    private LinkPhrases links;
     @SerializedName("scenario")
     private Scenario scenario;
     @SerializedName("numberOfSuites")